Change --dir to be the root directory
Previously, we used something like:
  _data/images/conf.json
  _data/images/image1.img
  _data/images/image2.img
  _data/kernels/conf.json
  _data/kernels/git
  _data/kernels/bpf-next

This is not very convinient because the --dir argument will change
depending on whether we using images or kernels commands.

This patch changes the commands so that --dir is the parent directory.
It also changes the configuration files to be saved there.

So now files look like:
  _data/images.json
  _data/images/image1.img
  _data/images/image2.img
  _data/kernels.json
  _data/kernels/git
  _data/kernels/bpf-next

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>