runner: Added VM runner
This commit add the `lvh run` command, which is a generalized version
of the Tetragon runner. It is essentially a small wrapper around qemu
to aid in setting the correct attributes to make it run. It modifies the
base image in its current iteration to add test scripts and systemd
services.

HOWTO run it:

    mkdir images
    docker run -v $(pwd)/images:/mnt/images \
        quay.io/lvh-images/root-images:latest \
        cp /data/images/kind.qcow2.zst /mnt/images
    cd images
    zstd -d kind.qcow2.zst && cd ..
    go run ./cmd/lvh run --image images/cilium.qcow2 \
        --host-mount ~/sandbox/gopath/src/github.com/cilium/ \
        --daemonize -p 2222:22
    ssh -p 2222 -o "StrictHostKeyChecking=no" root@localhost
    # hack hack

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>
Signed-off-by: Mark Pashmfouroush <mark@markpash.me>
Signed-off-by: Martynas Pumputis <m@lambda.lt>