images: build: add a chdir step
Add a step that changes the current directory to the main configuration
directory (i.e., the one provided with --dir). This allows actions to
refer to files inside this directory by relative paths. If we ever
needed it, we can make the dir configurable using an option.

For this to work, we also now pass the base dir as an absolute path.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>