action: Optionally set DNS resolver
The LVH VM images are built on Azure, and their /etc/resolv.conf point
to an Azure DNS resolvers. The DNS resolution works fine if running on
GHA runners (=Azure VMs).

However, this falls short once running on GKE, which we do via the poor
man's ephemeral GH action runners.

Fix this by allowing users to specify dns-resolver addr to be used in
the VMs.

Signed-off-by: Martynas Pumputis <m@lambda.lt>