Add renovate configuration
Renovate is a bot that allows to update dependencies in the repository.

The configuration is based on the one from the cilium/cilium repo:
https://github.com/cilium/cilium/blob/a42b70b55ef0fbe2e621b19db11d8d4c465cd85b/.github/renovate.json5
For future reference the documentation is available in:
- https://github.com/renovatebot/tutorial
- https://docs.renovatebot.com/configuration-options

Also switch all GH action workflows to pinned digests and version
comments, so renovate will properly update them.

Signed-off-by: Tobias Klauser <tobias@cilium.io>