action: add watch on console log file for sshd status
This commit updates the GHA action to watch for sshd status within the
VM's console log file. Success and failure messages for sshd are
continually searched for in the console log file over the course of 30s.
If a failure is observed, then the step fails. If success is observed,
then the step moves on to attempting to ssh into the VM. Continuing to
check if ssh access is available is important, because at the end of the
day, this is the functionality that is needed to continue. Parsing
through the console log file is an added step to help gain insight into
what is going on in the VM.

Regex match-alls is used for white-space in-between keywords when
searching, as console color characters may be printed between.
Additionally, '\.service' is not appended to the end of 'ssh' in these
search strings, as sometimes this part of the string gets truncated in
the log file.

Ref: cilium/cilium#26012

Signed-off-by: Ryan Drew <ryan.drew@isovalent.com>