kernels: add a raw_configure command
lvh assumes that kernels can be build by fetching them with git. This,
however, is not always possible. centos/rhel kernels are an example of
this. It would still be nice to use lvh to configure these kernels.

This patch adds a raw_configure command that allows just that.

The command accepts a source directory for the kernel, and an optional
kernel name. The latter is used for kernel-specific options as defined
in the configuration.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>