Add RELEASE.md
In particular, I'd like to make sure the default version in action.yaml
gets updated before pushing the release tag so that the default version
matches the action version. Right now the default version is always one
version behind, which is mildly inconvenient.

Sorry for all the whitespace changes, it should be a one time thing for
using yq to update action.yaml.

Signed-off-by: Michi Mutsuzaki <michi@isovalent.com>