gha: remove fragile SSH startup check
Currently, the LVH GH action verifies the successful start of the
SSH server by searching the console logoutput for the corresponding
log lines.

This mechanism is very fragile and leads to cases where a successful
SSH server start isn't detected properly. E.g. in cases where the logoutput
by qemu is interleaved.

```
[  OK  ] Started     4.887573] e2scrub_all (319) used greatest stack depth: 12416 bytes left
1;39mssh.service - OpenBSD Secure Shell server.
```

Therefore, this commit removes the SSH server startup check completely.
Leaving the check for a successful SSH connect to be the only check.

To respect the SSH server startup time, the default wait retries for the SSH
connection check gets increased from 5 to 300 (the time that we potentially
waited for the SSH startup in the log check).

Signed-off-by: Marco Hofstetter <marco.hofstetter@isovalent.com>