action: fix image derive step with multiple dashes
An image name that contains multiple dashes breaks because sed greedily matches on the
initial `.*` pattern. To fix this, change the initial `.*` to `[^-]` so that it matches
until we encounter the first dash character.

As a concrete example, prior to this change, `kind:6.6-20240707.080256-amd64` would
resolve to `kind_6.6-20240707.080256.qcow2` instead of `kind_6.6.qcow2` as expected. This
essentially means that all image names with two `-` characters are broken in this way.

Signed-off-by: William Findlay <will@isovalent.com>