action.yaml: Add support for LVH VM arch
This commits add a support for specifying different VM archs in the GH
action ("amd64" or "arm64").

To support different archs, we need to change the caching key, as the
multi-arch VM images use the same tag.

The images are stored in $images-folder-parent/images/$arch/images. The
"images" postfix is due to the "lvh images pull" automatically
preappending it to the $dir specified via "--dir $dir".

Signed-off-by: Martynas Pumputis <martynas@isovalent.com>